refactor: Betting, BettingTable domain으로 이동
