+++ src/test/java/blackjack/BettingTableTest.java
@@ -8,6 +8,8 @@ import java.util.Map;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import blackjack.domain.Betting;
+import blackjack.domain.BettingTable;
 import blackjack.domain.PlayRecord;
 import blackjack.domain.participant.Name;
