+++ src/test/java/blackjack/BettingTest.java
@@ -7,6 +7,7 @@ import java.util.Map;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import blackjack.domain.Betting;
 import blackjack.domain.PlayRecord;
 import blackjack.domain.participant.Name;
